requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.12.0
django-cpf==0.1.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
environ==1.0
Faker==4.14.2
filelock==3.7.0
gunicorn==20.1.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.6.0
itemloaders==1.0.4
itypes==1.2.0
jmespath==1.0.0
lockfile==0.12.2
lxml==4.8.0
Markdown==3.3.7
msgpack==1.0.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
node==1.0
numpy==1.22.4
odict==1.8.1
packaging==21.3
pandas==1.4.2
parsel==1.6.0
Pillow==9.1.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
plumber==1.7
poetry-core==1.0.8
Protego==0.2.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pylev==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
shellingham==1.4.0
six==1.15.0
sqlparse==0.4.2
text-unidecode==1.3
TgCrypto==1.2.3
tldextract==3.3.0
tomli==2.0.1
tomlkit==0.11.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
validate-docbr==1.8.2
w3lib==1.22.0
webencodings==0.5.1
whitenoise==6.1.0
zipp==3.7.0
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.4
zope.proxy==4.5.0
